+++ anyblok/test_bloks/test_blok8/declarations.py
@@ -16,7 +16,7 @@ Model = Declarations.Model
 @register(Model)
 class Test:
 
-    test2 = Integer(foreign_key=(Model.Test2, 'id'))
+    test2 = Integer(foreign_key=Model.Test2.use('id'))
 
 Declarations.AuthorizationBinding(Declarations.Model.Test,
                                   TestRuleOne(),