+++ anyblok/tests/test_one2one.py
@@ -164,9 +164,9 @@ class TestOne2One(DBTestCase):
 
                 id = Integer(primary_key=True)
                 test_id = Integer(
-                    foreign_key=(Model.Test, 'id'), nullable=False)
+                    foreign_key=Model.Test.use('id'), nullable=False)
                 test_id2 = String(
-                    foreign_key=(Model.Test, 'id2'), nullable=False)
+                    foreign_key=Model.Test.use('id2'), nullable=False)
                 test = One2One(model=Model.Test,
                                remote_columns=('id', 'id2'),
                                column_names=('test_id', 'test_id2'),
@@ -193,9 +193,9 @@ class TestOne2One(DBTestCase):
 
                 id = Integer(primary_key=True)
                 test_id = Integer(
-                    foreign_key=(Model.Test, 'id'), nullable=False)
+                    foreign_key=Model.Test.use('id'), nullable=False)
                 test_id2 = String(
-                    foreign_key=(Model.Test, 'id2'), nullable=False)
+                    foreign_key=Model.Test.use('id2'), nullable=False)
                 test = One2One(model=Model.Test,
                                remote_columns=('id', 'id2'),
                                column_names=('test_id', 'test_id2'),
@@ -222,9 +222,9 @@ class TestOne2One(DBTestCase):
 
                 id = Integer(primary_key=True)
                 test_id = Integer(
-                    foreign_key=(Model.Test, 'id'), nullable=False)
+                    foreign_key=Model.Test.use('id'), nullable=False)
                 test_id2 = String(
-                    foreign_key=(Model.Test, 'id2'), nullable=False)
+                    foreign_key=Model.Test.use('id2'), nullable=False)
                 test = One2One(model=Model.Test, backref="test2")
 
         registry = self.init_registry(add_in_registry)
@@ -248,9 +248,9 @@ class TestOne2One(DBTestCase):
 
                 id = Integer(primary_key=True)
                 other_test_id = Integer(
-                    foreign_key=(Model.Test, 'id'), nullable=False)
+                    foreign_key=Model.Test.use('id'), nullable=False)
                 other_test_id2 = String(
-                    foreign_key=(Model.Test, 'id2'), nullable=False)
+                    foreign_key=Model.Test.use('id2'), nullable=False)
                 test = One2One(model=Model.Test, backref="test2")
 
         registry = self.init_registry(add_in_registry)